ECDSASignature refactoring

+++ src/test/scala/io/iohk/ethereum/network/AuthInitiateMessageSpec.scala
@@ -23,7 +23,7 @@ class AuthInitiateMessageSpec extends FlatSpec with Matchers {
 
     val nonce = ByteUtils.randomBytes(AuthHandshaker.NonceSize)
 
-    val signature = ECDSASignature(new BigInteger("123"), new BigInteger("456"), 0)
+    val signature = ECDSASignature(new BigInteger("123"), new BigInteger("456"), 0.toByte)
 
     val msg = AuthInitiateMessage(
       signature,
@@ -42,7 +42,7 @@ class AuthInitiateMessageSpec extends FlatSpec with Matchers {
       signature = ECDSASignature(
         r = new BigInteger("81870901931874412952660009205824222047471672340278145383000560930072854380569"),
         s = new BigInteger("27900842753040147848386185004093503271309114686926362065982995477587410195214"),
-        v = 27),
+        v = 27.toByte),
       ephemeralPublicHash = ByteString(Hex.decode("F30BAC135324F493AEEAF4C9B48DF0F1F9A28A3DEDF9B1D85AF45A27F7B6F054")),
       publicKey = curve.getCurve.decodePoint(
         Hex.decode("046F8A80D14311C39F35F516FA664DEAAAA13E85B2F7493F37F6144D86991EC012937307647BD3B9A82ABE2974E1407241D54947BBB39763A4CAC9F77166AD92A0")),
